+++ src/widgets/numeric-selector/numeric-selector.js
@@ -37,6 +37,7 @@ const usage = `Usage: numericSelector({
 /**
  * @typedef {Object} NumericOption
  * @property {number} value The numerical value to refine with.
+ * If the value is `undefined` or `"undefined"`, the option resets the filter.
  * @property {string} label Label to display in the option.
  */
 
@@ -73,13 +74,16 @@ const usage = `Usage: numericSelector({
  * @example
  * search.addWidget(
  *   instantsearch.widgets.numericSelector({
- *     container: '#popularity-selector',
- *     attributeName: 'popularity',
- *     operator: '>=',
+ *     container: '#rating-selector',
+ *     attributeName: 'rating',
+ *     operator: '=',
  *     options: [
- *       {label: 'Top 10', value: 9900},
- *       {label: 'Top 100', value: 9800},
- *       {label: 'Top 500', value: 9700}
+ *       {label: 'All products'},
+ *       {label: 'Only 5 star products', value: 5},
+ *       {label: 'Only 4 star products', value: 4},
+ *       {label: 'Only 3 star products', value: 3},
+ *       {label: 'Only 2 star products', value: 2},
+ *       {label: 'Only 1 star products', value: 1},
  *     ]
  *   })
  * );